Renamed Frequency Switch files for consistency
