+++ handlers/product-switch-api/src/frequencySwitchSchemas.ts
