+++ handlers/product-switch-api/src/index.ts
@@ -3,9 +3,9 @@ import type { Stage } from '@modules/stage';
 import type { Handler } from 'aws-lambda';
 import dayjs from 'dayjs';
 import { z } from 'zod';
-import { frequencySwitchHandler } from './frequencyChange';
-import type { FrequencySwitchRequestBody } from './frequencySchemas';
-import { frequencySwitchRequestSchema } from './frequencySchemas';
+import { frequencySwitchHandler } from './frequencySwitchEndpoint';
+import type { FrequencySwitchRequestBody } from './frequencySwitchSchemas';
+import { frequencySwitchRequestSchema } from './frequencySwitchSchemas';
 import { contributionToSupporterPlusEndpoint } from './productSwitchEndpoint';
 import type { ProductSwitchRequestBody } from './schemas';
 import { productSwitchRequestSchema } from './schemas';
